Quote page now server component
